Rename actions.py to services.py
