+++ tests/components/home_connect/snapshots/test_services.ambr
